Time: 42 ms (16.13%), Space: 55.5 MB (42.36%) - LeetHub

+++ 0973-k-closest-points-to-origin/0973-k-closest-points-to-origin.java
@@ -1,8 +1,8 @@
 class triplet implements Comparable<triplet>{
     int x ;
     int y;
-    double dist;
-    triplet(int x,int y,double dist){
+    int dist;
+    triplet(int x,int y,int dist){
         this.x=x;
         this.y=y;
         this.dist=dist;
@@ -16,7 +16,7 @@ class Solution {
         int n = points.length;
         PriorityQueue<triplet> pq = new PriorityQueue<>();
         for(int i=0;i<n;i++){
-            double d = ((points[i][0]*points[i][0])+(points[i][1]*points[i][1]));
+            int d = ((points[i][0]*points[i][0])+(points[i][1]*points[i][1]));
             pq.add(new triplet(points[i][0],points[i][1],d));
         }
         while(pq.size()>k){
